+++ changelog.md
@@ -11,6 +11,7 @@ Codion Change Log
 - FilteredTable, Metal Look and Feel related bug fixed, searchField minimum width also used as preferred width.
 - HintTextField.hintText() renamed hint(), refactoring with minor improvements.
 - TextFieldBuilder.hintText() renamed hint().
+- DefaultListBuilder bug fixed, visibleRowCount was 0 by default.
 ### is.codion.swing.framework.ui
 - EntityPanel.icon() added, used when panel is displayed in windows or tabbed panes.
 - EntityPanel.Builder.icon() added.